Rework script to remove deprecation warnings
I'm not familiar with CoffeeScript (why fucking coffeescript gdi)
and for some reason refering to a function I'm trying to export (e.g.
launch) was throwing an error. So I just shoved the whole function into
the love-launcher:launch command.

It works and it removes the warnings. That's all that matters.